+++ src/main/java/xyz/zcraft/studios/zffa/gui/GuiType.java
@@ -2,13 +2,15 @@ package xyz.zcraft.studios.zffa.gui;
 
 public enum GuiType {
     MAIN,
+    MANAGEMENT,
+    KIT_EDITOR,
     KIT_SELECTOR,
     FFA_ARENAS,
     STATS,
     LEADERBOARD,
     PARTY,
-    PLAYER_MENU
-    ,DUEL_SELECTOR,
+    PLAYER_MENU,
+    DUEL_SELECTOR,
     RANKS,
     COSMETICS
 }